splits poster bookmarking into its own component
