+++ src/components/templates/PosterHall/components/PosterBookmark/index.ts
@@ -1,0 +1,1 @@
+export { PosterBookmark } from "./PosterBookmark";
